+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 Observes [Semantic Versioning](https://semver.org/spec/v2.0.0.html) standard and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) convention.
 
+## [3.0.0] - 06-26-2024
+
+### Added
+
+- support for python 3.12
+
+### Changed
+
+- update `DiscordChatExporter` to v2.43.3
+- update and pinned `openai` to v1.35.5
+- update reduction method to use an offset, a selection mode of either sequential or distributed, and a flag to iterate in reverse to reduce the line selection
+- update some default CLI command argument values
+- update some CLI help descriptions
+- major refactors
+
+### Removed
+
+- support for python 3.8
+
+- support for python 3.12
+
 ## [2.0.1] - 06-15-2023
 
 ### Changed
@@ -56,6 +77,7 @@ Observes [Semantic Versioning](https://semver.org/spec/v2.0.0.html) standard and
 
 - switched to `pathlib` for file path parsing
 
+[3.0.0]: https://github.com/A-Baji/discordAI-modelizer/compare/2.0.1...3.0.0
 [2.0.1]: https://github.com/A-Baji/discordAI-modelizer/compare/1.2.2...2.0.1
 [2.0.0]: https://github.com/A-Baji/discordAI-modelizer/compare/1.2.2...2.0.0
 [1.2.2]: https://github.com/A-Baji/discordAI-modelizer/compare/1.2.1...1.2.2